Reduce code duplication
The new file.js display used its own function to request derived-from files from the server as a GET request. This nearly duplicated existing code in filegallery.js, so moved this code to objectutils.js and both filegallery.js and file.js import it. Redmine #4650
